Factorize deployDestructiveChanges
