+++ src/commands/hardis/project/deploy/sources/dx.ts
@@ -5,11 +5,9 @@ import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import * as c from 'chalk';
 import * as fs from 'fs-extra';
-import * as os from 'os';
-import * as path from 'path';
 import { MetadataUtils } from '../../../../../common/metadata-utils';
-import { execCommand } from '../../../../../common/utils';
-import { CONSTANTS, getConfig } from '../../../../../config';
+import { execCommand, uxLog } from '../../../../../common/utils';
+import { getConfig } from '../../../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -57,7 +55,7 @@ export default class DxSources extends SfdxCommand {
   public async run(): Promise<AnyJson> {
     const check = this.flags.check || false;
     const testlevel = this.flags.testlevel || 'RunLocalTests';
-    const debug = this.flags.debug || false;
+    const debugMode = this.flags.debug || false;
     this.configInfo = await getConfig('branch');
 
     // Install packages
@@ -71,34 +69,9 @@ export default class DxSources extends SfdxCommand {
         (fs.existsSync('./manifest/destructiveChanges.xml')) ? './manifest/destructiveChanges.xml' :
         './config/destructiveChanges.xml';
     if (fs.existsSync(packageDeletedXmlFile)) {
-      // Create empty deployment file because of sfdx limitation
-      // cf https://gist.github.com/benahm/b590ecf575ff3c42265425233a2d727e
-      this.ux.log(`[sfdx-hardis] Deploying destructive changes from file ${path.resolve(packageDeletedXmlFile)}`);
-      const tmpDir = path.join(os.tmpdir(), 'sfdx-hardis-' + parseFloat(Math.random().toString()));
-      await fs.ensureDir(tmpDir);
-      const emptyPackageXmlFile = path.join(tmpDir, 'package.xml');
-      await fs.writeFile(emptyPackageXmlFile,
-        `<?xml version="1.0" encoding="UTF-8"?>
-        <Package xmlns="http://soap.sforce.com/2006/04/metadata">
-          <version>${CONSTANTS.API_VERSION}</version>
-        </Package>`, 'utf8');
-      await fs.copy(packageDeletedXmlFile, path.join(tmpDir, 'destructiveChanges.xml'));
-      const deployDelete = `sfdx force:mdapi:deploy -d ${tmpDir}` +
-        ' --wait 60' +
-        ' --testlevel NoTestRun' +
-        ' --ignorewarnings' + // So it does not fail in case metadata is already deleted
-        (check ? ' --checkonly' : '') +
-        (debug ? ' --verbose' : '');
-      const deployDeleteRes = await execCommand(deployDelete, this, { output: true, debug, fail: true });
-      await fs.remove(tmpDir);
-      let deleteMsg = '';
-      if (deployDeleteRes.status === 0) {
-        deleteMsg = '[sfdx-hardis] Successfully deployed destructive changes to Salesforce org';
-        this.ux.log(c.green(deleteMsg));
-      } else {
-        deleteMsg = '[sfdx-hardis] Unable to deploy destructive changes to Salesforce org';
-        this.ux.log(c.red(deployDeleteRes.errorMessage));
-      }
+      await MetadataUtils.deployDestructiveChanges(packageDeletedXmlFile, {debug: debugMode, check}, this);
+    } else {
+      uxLog( this, 'No destructivePackage.Xml found so no destructive deployment has been performed' );
     }
 
     // Deploy sources
@@ -111,8 +84,8 @@ export default class DxSources extends SfdxCommand {
       ' --wait 60' +
       ` --testlevel ${testlevel}` +
       (check ? ' --checkonly' : '') +
-      (debug ? ' --verbose' : '');
-    const deployRes = await execCommand(deployCommand, this, { output: true, debug, fail: true });
+      (debugMode ? ' --verbose' : '');
+    const deployRes = await execCommand(deployCommand, this, { output: true, debug: debugMode, fail: true });
     let message = '';
     if (deployRes.status === 0) {
       message = '[sfdx-hardis] Successfully deployed sfdx project sources to Salesforce org';
