+++ src/commands/hardis/project/deploy/sources/metadata.ts
@@ -9,7 +9,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { MetadataUtils } from '../../../../../common/metadata-utils';
 import { checkSfdxPlugin, execCommand, uxLog } from '../../../../../common/utils';
-import { CONSTANTS, getConfig } from '../../../../../config';
+import { getConfig } from '../../../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -90,7 +90,7 @@ export default class DxSources extends SfdxCommand {
         await MetadataUtils.installPackagesOnOrg(packages, null, this);
     }
 
-    let destructiveProcessed = false;
+    const destructiveProcessed = false;
     let deployProcessed = false;
 
     // Deploy destructive changes
@@ -104,60 +104,9 @@ export default class DxSources extends SfdxCommand {
         ? './destructiveChanges.xml'
         : './config/destructiveChanges.xml';
     if (fs.existsSync(packageDeletedXmlFile)) {
-      // Create empty deployment file because of sfdx limitation
-      // cf https://gist.github.com/benahm/b590ecf575ff3c42265425233a2d727e
-      this.ux.log(
-        `[sfdx-hardis] Deploying destructive changes from file ${path.resolve(
-          packageDeletedXmlFile
-        )}`
-      );
-      const tmpDir = path.join(
-        os.tmpdir(),
-        'sfdx-hardis-' + parseFloat(Math.random().toString())
-      );
-      await fs.ensureDir(tmpDir);
-      const emptyPackageXmlFile = path.join(tmpDir, 'package.xml');
-      await fs.writeFile(
-        emptyPackageXmlFile,
-        `<?xml version="1.0" encoding="UTF-8"?>
-        <Package xmlns="http://soap.sforce.com/2006/04/metadata">
-          <version>${CONSTANTS.API_VERSION}</version>
-        </Package>`,
-        'utf8'
-      );
-      await fs.copy(
-        packageDeletedXmlFile,
-        path.join(tmpDir, 'destructiveChanges.xml')
-      );
-      const deployDelete =
-        `sfdx force:mdapi:deploy -d ${tmpDir}` +
-        ' --wait 60' +
-        ' --testlevel NoTestRun' +
-        ' --ignorewarnings' + // So it does not fail in case metadata is already deleted
-        (check ? ' --checkonly' : '') +
-        (debugMode ? ' --verbose' : '');
-      const deployDeleteRes = await execCommand(deployDelete, this, {
-        output: true,
-        debugMode,
-        fail: true
-      });
-      await fs.remove(tmpDir);
-      let deleteMsg = '';
-      if (deployDeleteRes.status === 0) {
-        destructiveProcessed = true;
-        deleteMsg =
-          '[sfdx-hardis] Successfully deployed destructive changes to Salesforce org';
-        this.ux.log(c.green(deleteMsg));
-      } else {
-        deleteMsg =
-          '[sfdx-hardis] Unable to deploy destructive changes to Salesforce org';
-        this.ux.log(c.red(deployDeleteRes.errorMessage));
-      }
+        await MetadataUtils.deployDestructiveChanges(packageDeletedXmlFile, {debug: debugMode, check}, this);
     } else {
-      uxLog(
-        this,
-        'No destructivePackageXml found so no destructive deployment has been performed'
-      );
+      uxLog( this, 'No destructivePackage.Xml found so no destructive deployment has been performed' );
     }
 
     // Deploy sources
